% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecolors.R
\name{treecolors}
\alias{treecolors}
\title{Create color palette based on hierarchical structure}
\usage{
treecolors(...)
}
\arguments{
\item{...}{not used}
}
\description{
Create color palette based on hierarchical structure
}

